Refactorized view models
Moved to Infrasctructure/ReadModel/View
