+++ spec/Application/Role/Query/GetRoleHandlerSpec.php
@@ -16,8 +16,8 @@ namespace spec\AulaSoftwareLibre\Iam\Application\Role\Query;
 use AulaSoftwareLibre\Iam\Application\Role\Query\GetRole;
 use AulaSoftwareLibre\Iam\Application\Role\Query\GetRoleHandler;
 use AulaSoftwareLibre\Iam\Domain\Role\Model\RoleId;
-use AulaSoftwareLibre\Iam\Infrastructure\ReadModel\Role\View\RoleView;
 use AulaSoftwareLibre\Iam\Infrastructure\ReadModel\Repository\RoleViews;
+use AulaSoftwareLibre\Iam\Infrastructure\ReadModel\View\RoleView;
 use PhpSpec\ObjectBehavior;
 use React\Promise\Deferred;
 use Tests\Spec\Fixtures;
