+++ spec/Application/Role/Query/GetRolesHandlerSpec.php
@@ -16,8 +16,8 @@ namespace spec\AulaSoftwareLibre\Iam\Application\Role\Query;
 use AulaSoftwareLibre\Iam\Application\Role\Query\GetRoles;
 use AulaSoftwareLibre\Iam\Application\Role\Query\GetRolesHandler;
 use AulaSoftwareLibre\Iam\Domain\Scope\Model\ScopeId;
-use AulaSoftwareLibre\Iam\Infrastructure\ReadModel\Role\View\RoleView;
 use AulaSoftwareLibre\Iam\Infrastructure\ReadModel\Repository\RoleViews;
+use AulaSoftwareLibre\Iam\Infrastructure\ReadModel\View\RoleView;
 use PhpSpec\ObjectBehavior;
 use React\Promise\Deferred;
 use Tests\Spec\Fixtures;
