+++ src/Infrastructure/ReadModel/View/RoleView.php
@@ -11,11 +11,9 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace AulaSoftwareLibre\Iam\Infrastructure\ReadModel\Scope\View;
+namespace AulaSoftwareLibre\Iam\Infrastructure\ReadModel\View;
 
-use AulaSoftwareLibre\Iam\Infrastructure\ReadModel\Role\View\RoleView;
-
-class ScopeView
+class RoleView
 {
     /**
      * @var string
@@ -24,21 +22,24 @@ class ScopeView
     /**
      * @var string
      */
-    private $name;
+    private $scopeId;
     /**
      * @var string
      */
-    private $shortName;
-    /**
-     * @var RoleView[]|array
-     */
-    private $roles;
+    private $name;
 
-    public function __construct(string $id, string $name, string $shortName)
+    /**
+     * RoleView constructor.
+     *
+     * @param string $id
+     * @param string $scopeId
+     * @param string $name
+     */
+    public function __construct(string $id, string $scopeId, string $name)
     {
         $this->id = $id;
+        $this->scopeId = $scopeId;
         $this->name = $name;
-        $this->shortName = $shortName;
     }
 
     /**
@@ -52,24 +53,16 @@ class ScopeView
     /**
      * @return string
      */
-    public function getName(): string
+    public function getScopeId(): string
     {
-        return $this->name;
-    }
-
-    /**
-     * @param string $name
-     */
-    public function setName(string $name): void
-    {
-        $this->name = $name;
+        return $this->scopeId;
     }
 
     /**
      * @return string
      */
-    public function getShortName(): string
+    public function getName(): string
     {
-        return $this->shortName;
+        return $this->name;
     }
 }
