+++ src/Infrastructure/ReadModel/View/UserView.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace AulaSoftwareLibre\Iam\Infrastructure\ReadModel\User\View;
+namespace AulaSoftwareLibre\Iam\Infrastructure\ReadModel\View;
 
 class UserView
 {
