+++ src/Infrastructure/Resources/config/doctrine/UserView.orm.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping">
     
-    <entity name="AulaSoftwareLibre\Iam\Infrastructure\ReadModel\User\View\UserView" table="user_view">
+    <entity name="AulaSoftwareLibre\Iam\Infrastructure\ReadModel\View\UserView" table="user_view">
         <id name="id" column="id" type="string" />
         <field name="username" type="string" unique="true" />
         <field name="email" type="string" unique="true" />
